requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
pytube==15.0.0
youtube-search-python==1.6.6
redis==latest
